collapse different presentation formats into single file
