--- internal/v2client/json_presenter.go
@@ -1,18 +1,0 @@
-package v2client
-
-import (
-	"encoding/json"
-	"fmt"
-
-	log "github.com/sirupsen/logrus"
-)
-
-func (p *Presenter) asJSON() {
-	j, err := json.Marshal(p.ServiceReport)
-	if err != nil {
-		log.Fatalln(err)
-	}
-
-	// TODO maybe there's a cleaner way of outputting this than just a println
-	fmt.Println(string(j))
-}
