+++ internal/v2client/presenter.go
@@ -1,7 +1,13 @@
 package v2client
 
 import (
+	"encoding/json"
+	"fmt"
+	"os"
+
 	log "github.com/sirupsen/logrus"
+
+	"github.com/olekukonko/tablewriter"
 )
 
 // Presenter -
@@ -26,4 +32,34 @@ func (p *Presenter) Render() {
 			log.Debugf("unknown format [%s], using default\n", p.Format)
 		}
 	}
+}
+
+func (p *Presenter) asTable() {
+	table := tablewriter.NewWriter(os.Stdout)
+	table.SetHeader([]string{
+		"service_guid",
+		"service",
+		"org",
+		"space",
+	})
+
+	for _, report := range p.ServiceReport {
+		table.Append([]string{
+			report.Service.GUID,
+			report.Service.Name,
+			report.Organization.Name,
+			report.Space.Name,
+		})
+	}
+
+	table.Render()
+}
+
+func (p *Presenter) asJSON() {
+	j, err := json.Marshal(p.ServiceReport)
+	if err != nil {
+		log.Fatalln(err)
+	}
+
+	fmt.Println(string(j))
 }
