--- internal/v2client/table_presenter.go
@@ -1,29 +1,0 @@
-package v2client
-
-import (
-	"os"
-
-	"github.com/olekukonko/tablewriter"
-)
-
-func (p *Presenter) asTable() {
-	table := tablewriter.NewWriter(os.Stdout)
-	table.SetHeader([]string{
-		"service_guid",
-		"service",
-		"org",
-		"space",
-	})
-
-	for _, report := range p.ServiceReport {
-		table.Append([]string{
-			report.Service.GUID,
-			report.Service.Name,
-			report.Organization.Name,
-			report.Space.Name,
-		})
-	}
-
-	table.Render()
-
-}
